Rename package to satokengen to make it clearer that it is for service account tokens
