+++ pkg/components/satokengen/apikeygen.go
@@ -1,4 +1,4 @@
-package apikeygenprefix
+package satokengen
 
 import (
 	"encoding/hex"
