+++ pkg/components/satokengen/cmd/main.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"os"
 	"strconv"
-
-	"github.com/grafana/grafana/pkg/components/apikeygenprefix"
 )
 
 // placeholder key generator
@@ -22,7 +20,7 @@ func main() {
 	}
 
 	for i := 0; i < numKeys; i++ {
-		key, err := apikeygenprefix.New("pl")
+		key, err := satokengen.New("pl")
 		if err != nil {
 			fmt.Println("ERROR: generating key failed:", err)
 			return
